--- .vscode/launch.json
@@ -1,12 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "msedge",
-      "request": "launch",
-      "name": "Launch Edge against localhost",
-      "url": "http://localhost:5500",
-      "webRoot": "${workspaceFolder}/dist"
-    }
-  ]
-}
