+++ README.md
@@ -10,5 +10,5 @@ Protection provided by [QuSecure](https://www.qusecure.com)
 Authentication provided by [iValt](https://www.ivalt.com)
 
 ## Packages
-- [`@trans-stat/biometrics`](https://github.com/trans-stat/tas-security/blob/main/packages/biometrics) - Node.js esm and common.js compatible wrapper around iValt's biometric authentication API
+- [`@trans-stat/node-mfa`](https://github.com/trans-stat/tas-security/blob/main/packages/node-mfa) - Node.js ESM and CommonJS compatible wrappers for a variety of multifactor authentication methods.
 - [`@trans-stat/auth-components`](https://github.com/trans-stat/tas-security/blob/main/packages/auth-components) - Web Components for user authentication UX patterns.
