+++ packages/node-mfa/src/index.ts
@@ -1,0 +1,1 @@
+export * from './ivalt/index.js';
