+++ packages/web-components/README.md
@@ -3,8 +3,8 @@ Web Components built with FAST for easy user authentication UX.
 
 > NOTE: This package is still early in development. Expect breaking API changes often until the package reaches a stable 1.0.0 release.
 
-> NOTE: This library is meant to be used in conjuction with the backend packages published from [tas-security](https://github.com/trans-stat/tas-security).
-While these frontend Web Components can be used without the backend packages, they will work best as a fullstack solution.
+> NOTE: This library is meant to be used in conjuction with the backend package published from [tas-security](https://github.com/trans-stat/tas-security).
+While these frontend Web Components can be used without the backend package, they will work best as a fullstack solution.
 
 ## installation
 ### npm
@@ -30,27 +30,27 @@ Add a `.vscode/settings.json` file to the root of your repository and set the ht
 }
 ```
 
-### `<biometric-card>`
-> NOTE: This element is intended to be used with the `@trans-stat/biometrics` package in the backend.
-These two packages create a fullstack wrapper for [iValt's](https://ivalt.com) biometrics API
+### `<mfa-card>`
+> NOTE: This element is intended to be used with the `@trans-stat/node-mfa` package in the backend.
+These two packages create a fullstack wrapper for [iValt's](https://ivalt.com) multifact authentication API
 1. Define the component with the platform
 ```ts
-import '@trans-stat/auth-components/biometric-card';
+import '@trans-stat/auth-components/mfa-card';
 ```
 2. Add the element to your markup.
 ```html
-<biometric-card id="bio-card"></biometric-card>
+<mfa-card id="mfa-card"></mfa-card>
 ```
 3. hook into the element to send its data to your backend.
 ```ts
-import type { BiometricCard } from '@trans-stat/auth-components';
+import type { MFACard } from '@trans-stat/auth-components';
 
-const bioCard = document.getElementById('bio-card') as BiometricCard;
+const mfaCard = document.getElementById('mfa-card') as MFACard;
 
 /**
 * The element provides two different hooks depending on your needs.
 * The first hook is a callback property that the element calls when the submit button is clicked.
-* The second hook is a custom event called `biometric-request` that gets emitted when the submit button is clicked.
+* The second hook is a custom `changed` event that gets emitted when the submit button is clicked.
 *
 * Both methods get passed the element's form data.
 * 
@@ -60,7 +60,7 @@ const bioCard = document.getElementById('bio-card') as BiometricCard;
 // Option 1. custom callback
 bioCard.callback = async (formData) => {
   const response = await fetch(
-    '/my-app-biometrics-endpoint',
+    '/my-app-mfa-endpoint',
     {
       method: 'POST',
       headers: {
@@ -72,9 +72,9 @@ bioCard.callback = async (formData) => {
 }
 
 // Option 2. event handler
-bioCard.addEventListener('biometric-request', async (event: CustomEvent<{ mobileNumber: string }>) => {
+bioCard.addEventListener('changed', async (event: CustomEvent<{ mobileNumber: string }>) => {
   const response = await fetch(
-    '/my-app-biometrics-endpoint',
+    '/my-app-mfa-endpoint',
     {
       method: 'POST',
       headers: {
