--- packages/web-components/src/biometric-card/index.ts
@@ -1,14 +1,0 @@
-import { BiometricCard } from './biometric-card.js';
-import { biometricCardTemplate } from './biometric-card.template.js';
-import { biometricCardStyles } from './biometric-card.styles.js';
-
-export type {
-  BiometricCardCallback,
-  BiometricCardTemplateOptions
-} from './biometric-card.options.js';
-
-export {
-  BiometricCard,
-  biometricCardTemplate,
-  biometricCardStyles,
-};
