+++ packages/web-components/src/mfa-card/index.ts
@@ -1,0 +1,14 @@
+import { MFACard } from './mfa-card.js';
+import { mfaCardTemplate } from './mfa-card.template.js';
+import { mfaCardStyles } from './mfa-card.styles.js';
+
+export type {
+  MFACardCallback,
+  MFACardTemplateOptions
+} from './mfa-card.options.js';
+
+export {
+  MFACard,
+  mfaCardTemplate,
+  mfaCardStyles,
+};
